added openmpi to docker image; using relative imports;

+++ scripts/colm_pipeline.py
@@ -1,10 +1,11 @@
-from download_raw_data import download_raw_data
-from correct_raw_data import correct_raw_data
-from create_precomputed_volume import create_precomputed_volume
-from correct_stitched_data import correct_stitched_data
-from stitching import run_terastitcher
-from util import S3Url, upload_file_to_s3, download_file_from_s3, download_terastitcher_files, tqdm_joblib, aws_cli
-from visualization import create_viz_link
+from .download_raw_data import download_raw_data
+from .correct_raw_data import correct_raw_data
+from .create_precomputed_volume import create_precomputed_volume
+from .correct_stitched_data import correct_stitched_data
+from .stitching import run_terastitcher
+from .util import S3Url, upload_file_to_s3, download_file_from_s3, download_terastitcher_files, tqdm_joblib, aws_cli
+from .visualization import create_viz_link
+
 import numpy as np
 from glob import glob
 from tqdm import tqdm
@@ -136,4 +137,5 @@ if __name__ == "__main__":
                     ) for f in directories_to_remove
                 )
             # make sure to delete mdata.bin from terastitcher
-            os.remove(f'{args.raw_data_path}/mdata.bin')
+            if os.path.exists(f'{args.raw_data_path}/mdata.bin'):
+                os.remove(f'{args.raw_data_path}/mdata.bin')
